+++ test/data_element_group_test.rb
@@ -1,15 +1,11 @@
 require "test_helper"
 
 class DataElementGroupTest < Minitest::Test
-
   def test_list_data_element_groups
-
     data_element_groups = Dhis2::DataElementGroup.list(fields: %w(id displayShortName displayName code), page_size: 1)
     assert_equal 1, data_element_groups.size
 
     data_element_group = data_element_groups.first
-
-    puts data_element_group.inspect
 
     refute_nil data_element_group.display_name
     refute_nil data_element_group.id
@@ -17,10 +13,12 @@ class DataElementGroupTest < Minitest::Test
 
   def test_create_data_element_groups
     groups = [
-      { name: "group_one", short_name: "gone" },
+      { name: "group_one", short_name: "groutp_one" },
       { name: "group_two", short_name: "gtwo" }
     ]
+
     status = Dhis2::DataElementGroup.create(groups)
+
     assert_equal true, status.success?
     assert_equal 2, status.total_imported
   end
@@ -34,5 +32,4 @@ class DataElementGroupTest < Minitest::Test
     refute_nil data_element_group.display_name
     refute_nil data_element_group.id
   end
-
 end
